Implementation LD
Inmplementation of LD 1-15, those that need complete rework are empty until done (2 or 3 out of the 15).
LDs that have the "Escape" goal have an asteroid put where the Escape tile should be as the Escape tile does not yet exist in the project.
All LDs are duplicate of the scene "LD Testing Ground" which itself is a copy of "TileScene" emptied of its ennemies and special tiles.
LDs are sorted by an order which may change soon, after discussing it with the other games designers monday morning.